code cleanup

+++ Test Game/DeviceContext.cpp
@@ -1,20 +1,19 @@
 #include "DeviceContext.h"
-#include "SwapChain.h"
-#include "VertexBuffer.h"
-#include "IndexBuffer.h"
-#include "ConstantBuffer.h"
-#include "VertexShader.h"
-#include "PixelShader.h"
-#include "Texture.h"
 #include <exception>
+#include "ConstantBuffer.h"
+#include "IndexBuffer.h"
+#include "PixelShader.h"
+#include "SwapChain.h"
+#include "Texture.h"
+#include "VertexBuffer.h"
+#include "VertexShader.h"
 
-DeviceContext::DeviceContext(ID3D11DeviceContext* device_context, RenderSystem * system) : m_system(system) ,m_device_context(device_context)
-{
-}
+DeviceContext::DeviceContext(ID3D11DeviceContext* device_context, RenderSystem * system) : m_system(system) ,m_device_context(device_context) {}
 
 void DeviceContext::clearRenderTargetColor(const SwapChainPtr& swap_chain, float red, float green, float blue, float alpha)
 {
 	FLOAT clear_color[] = {red,green,blue,alpha};
+	
 	m_device_context->ClearRenderTargetView(swap_chain->m_rtv, clear_color);
 	m_device_context->ClearDepthStencilView(swap_chain->m_dsv, D3D11_CLEAR_DEPTH | D3D11_CLEAR_STENCIL, 1, 0);
 	m_device_context->OMSetRenderTargets(1, &swap_chain->m_rtv, swap_chain->m_dsv);
@@ -42,7 +41,7 @@ void DeviceContext::drawTriangleList(UINT vertex_count, UINT start_vertex_index)
 void DeviceContext::drawIndexedTriangleList(UINT index_count, UINT start_vertex_index, UINT start_index_location)
 {
 	m_device_context->IASetPrimitiveTopology(D3D11_PRIMITIVE_TOPOLOGY_TRIANGLELIST);
-	m_device_context->DrawIndexed(index_count, start_index_location, start_vertex_index);
+	m_device_context->DrawIndexed(index_count, start_index_location, (INT)start_vertex_index);
 }
 
 void DeviceContext::drawTriangleStrip(UINT vertex_count, UINT start_vertex_index)
@@ -88,11 +87,13 @@ void DeviceContext::setTexture(const PixelShaderPtr& pixel_shader, const Texture
 {
 	ID3D11ShaderResourceView* list_res[32];
 	ID3D11SamplerState* list_sampler[32];
+	
 	for (unsigned int i = 0; i < num_textures; i++)
 	{
 		list_res[i] = texture[i]->m_shader_res_view;
 		list_sampler[i] = texture[i]->m_sampler_state;
 	}
+	
 	m_device_context->PSSetShaderResources(0, num_textures, list_res);
 	m_device_context->PSSetSamplers(0, num_textures, list_sampler);
 }
@@ -106,7 +107,6 @@ void DeviceContext::setConstantBuffer(const PixelShaderPtr & pixel_shader, const
 {
 	m_device_context->PSSetConstantBuffers(0, 1, &buffer->m_buffer);
 }
-
 
 DeviceContext::~DeviceContext()
 {
